+++ api/v1alpha1/doc.go
@@ -1,0 +1,5 @@
+// +k8s:openapi-gen=true
+// +k8s:deepcopy-gen=package,register
+
+// +groupName=monitoring.giantswarm.io
+package v1alpha1