+++ service/controller/key/key.go
@@ -3,22 +3,23 @@ package key
 import (
 	"fmt"
 
-	monitoringv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/monitoring/v1alpha1"
 	"github.com/giantswarm/microerror"
+
+	"github.com/giantswarm/silence-operator/api/v1alpha1"
 )
 
 const (
 	CreatedBy = "silence-operator"
 )
 
-func ToSilence(v interface{}) (monitoringv1alpha1.Silence, error) {
+func ToSilence(v interface{}) (v1alpha1.Silence, error) {
 	if v == nil {
-		return monitoringv1alpha1.Silence{}, microerror.Maskf(wrongTypeError, "expected non-nil, got %#v", v)
+		return v1alpha1.Silence{}, microerror.Maskf(wrongTypeError, "expected non-nil, got %#v", v)
 	}
 
-	p, ok := v.(*monitoringv1alpha1.Silence)
+	p, ok := v.(*v1alpha1.Silence)
 	if !ok {
-		return monitoringv1alpha1.Silence{}, microerror.Maskf(wrongTypeError, "expected %T, got %T", p, v)
+		return v1alpha1.Silence{}, microerror.Maskf(wrongTypeError, "expected %T, got %T", p, v)
 	}
 
 	c := p.DeepCopy()
@@ -26,6 +27,6 @@ func ToSilence(v interface{}) (monitoringv1alpha1.Silence, error) {
 	return *c, nil
 }
 
-func SilenceComment(silence monitoringv1alpha1.Silence) string {
+func SilenceComment(silence v1alpha1.Silence) string {
 	return fmt.Sprintf("%s-%s", CreatedBy, silence.Name)
 }